requirements.txt
aiodns==2.0.0
aiohttp==3.6.3
aiohttp-socks==0.4.1
APScheduler==3.6.3
async-timeout==3.0.1
attrs==20.2.0
beautifulsoup4==4.9.3
cchardet==2.1.6
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.1.1
dataclasses==0.6
decorator==4.4.2
elasticsearch==7.9.1
fake-useragent==0.1.11
geographiclib==1.50
geopy==2.0.0
googletransx==2.4.2
idna==2.10
multidict==4.7.6
numpy==1.19.2
pandas==1.1.3
pycares==3.1.1
pycparser==2.20
PySocks==1.7.1
python-dateutil==2.8.1
python-telegram-bot==13.0
pytz==2020.1
requests==2.24.0
schedule==0.6.0
six==1.15.0
soupsieve==2.0.1
tornado==6.0.4
tzlocal==2.1
urllib3==1.25.11
yarl==1.5.1
